refactor: add Textarea class
